+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Changelog
 
-- updated documentation (GettingStarted.mlx and README.md)
+- fixed buggy UI for R2022a and MATLAB online (fitter panel)